Add proof of concept for API keys.
Adds a getMyProjectOverviews() function to Thrift which expects an API key and returns only projects authored by the user associated with that API key.

Hardcodes the only API key as 1234567890abcdef and hardcodes it to the user test_user.

Makes the Python client library's create_syncer accept an optional api_key param.

Adds get_project_overviews() and get_my_project_overviews() methods to the Syncer to allow the simplest example of a client setting their API key when they create a syncer and letting the client library pass it along to the server as necessary.